/feature add Areas for Admin and User
