+++ ProjectWeb/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using ASP_NET_core_Project
+@using ASP_NET_core_Project.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
