--- ProjectWeb/Controllers/HomeController.cs
@@ -1,31 +1,0 @@
-using System.Diagnostics;
-using Microsoft.AspNetCore.Mvc;
-using ASP_NET_core_Project.Models;
-
-namespace ASP_NET_core_Project.Controllers;
-
-public class HomeController : Controller
-{
-    private readonly ILogger<HomeController> _logger;
-
-    public HomeController(ILogger<HomeController> logger)
-    {
-        _logger = logger;
-    }
-
-    public IActionResult Index()
-    {
-        return View();
-    }
-
-    public IActionResult Privacy()
-    {
-        return View();
-    }
-
-    [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-    public IActionResult Error()
-    {
-        return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-    }
-}
