+++ ProjectWeb/Views/Shared/_Layout.cshtml.css
@@ -7,9 +7,9 @@ a.navbar-brand {
   word-break: break-all;
 }
 
-a {
+/*a {
   color: #0077cc;
-}
+}*/
 
 .btn-primary {
   color: #fff;
